+++ FSolv/model/Contribuinte.cs
@@ -20,7 +20,7 @@ namespace FSolv.model
 {
     public class Contribuinte : IContribuinte
     {
-        public virtual int? Nif { get; set; }
+        public virtual long? Nif { get; set; }
         public virtual string Name { get; set; }
         public virtual string Morada { get; set; }
         public virtual List<IFatura> Faturas { get; set; }
