+++ FSolv/model/Produto.cs
@@ -8,8 +8,8 @@ namespace FSolv.model
     public class Produto : IProduto
     {
         public int? Sku { get; set; }
-        public double Valor { get; set; }
-        public double Iva { get; set; }
+        public decimal Valor { get; set; }
+        public decimal Iva { get; set; }
         public string Descricao { get; set; }
     }
 }
