+++ Interfaces/model/IItem.cs
@@ -9,7 +9,7 @@ namespace Interfaces
     public interface IItem
     {
         int? Id { get; set; }
-        double Desconto { get; set; }
+        decimal Desconto { get; set; }
         int Qnt { get; set; }
         IProduto Produto { get; set; }
         IFatura Fatura { get; set; }
